#19 myPage 중첩 리사이클러뷰 연결
