+++ app/src/main/java/org/sopt/appzam/nobar_android/presentation/main/HomeStatePagerAdapter.kt
@@ -4,7 +4,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import org.sopt.appzam.nobar_android.presentation.main.home.HomeFragment
-import org.sopt.appzam.nobar_android.presentation.main.mypage.MypageFragment
+import org.sopt.appzam.nobar_android.presentation.main.mypage.MyPageFragment
 import org.sopt.appzam.nobar_android.presentation.main.record.RecordFragment
 import org.sopt.appzam.nobar_android.presentation.main.search.SearchFragment
 
@@ -18,7 +18,7 @@ class HomeStatePagerAdapter(fragmentActivity: FragmentActivity) :
             0 -> HomeFragment()
             1 -> SearchFragment()
             2 -> RecordFragment()
-            3 -> MypageFragment()
+            3 -> MyPageFragment()
             else -> error("position $position is cannot exist")
         }
     }
