+++ app/src/main/java/org/sopt/appzam/nobar_android/presentation/main/mypage/adapter/TastingNoteAdapter.kt
@@ -18,6 +18,11 @@ class TastingNoteAdapter :
         RecyclerView.ViewHolder(binding.root) {
         fun onBind(data: MyPageTastingResponse) {
             binding.myPageTastingNoteItem = data
+            val tastingTagAdapter = TastingTagAdapter().apply {
+                tastingTagList.clear()
+                tastingTagList.addAll(data.tag)
+            }
+            binding.recyclerTag.adapter = tastingTagAdapter
         }
 
     }
