+++ app/src/main/res/layout/fragment_mypage.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.main.mypage.MypageFragment">
+        tools:context=".presentation.main.mypage.MyPageFragment">
         <TextView
             android:id="@+id/textTitle"
             android:layout_width="wrap_content"
