+++ app/src/main/res/layout/item_my_page_tasting_note.xml
@@ -58,6 +58,7 @@
             android:layout_marginTop="16dp" />
         
         <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recyclerTag"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:layout_constraintStart_toStartOf="parent"
