+++ app/src/main/res/layout/item_my_page_tasting_tag.xml
@@ -27,6 +27,7 @@
             app:layout_constraintBottom_toBottomOf="parent"/>
 
         <TextView
+            android:id="@+id/textTagName"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="다섯글자로"
